Move color structs to their own module
